backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tensorflow>=2.13.0
pydantic>=2.0.0
